Move logger to another file
